--- axum-handle-error-extract/CHANGELOG.md
@@ -1,14 +1,0 @@
-# Changelog
-
-All notable changes to this project will be documented in this file.
-
-The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
-and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
-
-# Unreleased
-
-- None
-
-# 0.1.0 (05. November, 2021)
-
-- Initial release.
