--- axum-handle-error-extract/Cargo.toml
@@ -1,22 +1,0 @@
-[package]
-name = "axum-handle-error-extract"
-version = "0.1.0"
-authors = ["David Pedersen <david.pdrsn@gmail.com>"]
-categories = ["asynchronous", "network-programming", "web-programming"]
-description = "Error handling layer for axum that supports extractors and async functions"
-edition = "2018"
-homepage = "https://github.com/tokio-rs/axum"
-keywords = ["http", "web", "framework"]
-license = "MIT"
-readme = "README.md"
-repository = "https://github.com/tokio-rs/axum"
-
-[dependencies]
-axum = { version = "0.3.2", path = "../axum" }
-tower-service = "0.3"
-tower-layer = "0.3"
-tower = { version = "0.4", features = ["util"] }
-pin-project-lite = "0.2"
-
-[dev-dependencies]
-tower = { version = "0.4", features = ["util", "timeout"] }
