--- axum-handle-error-extract/README.md
@@ -1,45 +1,0 @@
-# axum-handle-error-extract
-
-Error handling layer for axum that supports extractors and async functions
-
-[![Build status](https://github.com/tokio-rs/axum-handle-error-extract/actions/workflows/CI.yml/badge.svg?branch=main)](https://github.com/tokio-rs/axum-handle-error-extract/actions/workflows/CI.yml)
-[![Crates.io](https://img.shields.io/crates/v/axum-handle-error-extract)](https://crates.io/crates/axum-handle-error-extract)
-[![Documentation](https://docs.rs/axum-handle-error-extract/badge.svg)](https://docs.rs/axum-handle-error-extract)
-
-More information about this crate can be found in the [crate documentation][docs].
-
-## Safety
-
-This crate uses `#![forbid(unsafe_code)]` to ensure everything is implemented in
-100% safe Rust.
-
-## Minimum supported Rust version
-
-axum-handle-error-extract's MSRV is 1.54.
-
-## Getting Help
-
-You're also welcome to ask in the [Discord channel][chat] or open an [issue]
-with your question.
-
-## Contributing
-
-:balloon: Thanks for your help improving the project! We are so happy to have
-you! We have a [contributing guide][contributing] to help you get involved in the
-`axum` project.
-
-## License
-
-This project is licensed under the [MIT license][license].
-
-### Contribution
-
-Unless you explicitly state otherwise, any contribution intentionally submitted
-for inclusion in `axum` by you, shall be licensed as MIT, without any
-additional terms or conditions.
-
-[docs]: https://docs.rs/axum-handle-error-extract
-[contributing]: /CONTRIBUTING.md
-[chat]: https://discord.gg/tokio
-[issue]: https://github.com/tokio-rs/axum/issues/new
-[license]: /axum/LICENSE
