+++ axum/src/error_handling/mod.rs
@@ -1,15 +1,20 @@
 #![doc = include_str!("../docs/error_handling.md")]
 
-use crate::{body::BoxBody, response::IntoResponse, BoxError};
-use bytes::Bytes;
-use http::{Request, Response};
+use crate::{
+    body::{boxed, BoxBody, Bytes, Full, HttpBody},
+    extract::{FromRequest, RequestParts},
+    http::{Request, Response, StatusCode},
+    response::IntoResponse,
+    BoxError,
+};
 use std::{
     convert::Infallible,
     fmt,
+    future::Future,
     marker::PhantomData,
     task::{Context, Poll},
 };
-use tower::{util::Oneshot, ServiceExt as _};
+use tower::ServiceExt;
 use tower_layer::Layer;
 use tower_service::Service;
 
@@ -17,49 +22,34 @@ use tower_service::Service;
 /// that handles errors by converting them into responses.
 ///
 /// See [module docs](self) for more details on axum's error handling model.
-pub struct HandleErrorLayer<F, B> {
+pub struct HandleErrorLayer<F, T> {
     f: F,
-    _marker: PhantomData<fn() -> B>,
+    _extractor: PhantomData<fn() -> T>,
 }
 
-impl<F, B> HandleErrorLayer<F, B> {
+impl<F, T> HandleErrorLayer<F, T> {
     /// Create a new `HandleErrorLayer`.
     pub fn new(f: F) -> Self {
         Self {
             f,
-            _marker: PhantomData,
+            _extractor: PhantomData,
         }
     }
 }
 
-impl<F, B, S> Layer<S> for HandleErrorLayer<F, B>
-where
-    F: Clone,
-{
-    type Service = HandleError<S, F, B>;
-
-    fn layer(&self, inner: S) -> Self::Service {
-        HandleError {
-            inner,
-            f: self.f.clone(),
-            _marker: PhantomData,
-        }
-    }
-}
-
-impl<F, B> Clone for HandleErrorLayer<F, B>
+impl<F, T> Clone for HandleErrorLayer<F, T>
 where
     F: Clone,
 {
     fn clone(&self) -> Self {
         Self {
             f: self.f.clone(),
-            _marker: PhantomData,
+            _extractor: PhantomData,
         }
     }
 }
 
-impl<F, B> fmt::Debug for HandleErrorLayer<F, B> {
+impl<F, E> fmt::Debug for HandleErrorLayer<F, E> {
     fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
         f.debug_struct("HandleErrorLayer")
             .field("f", &format_args!("{}", std::any::type_name::<F>()))
@@ -67,37 +57,52 @@ impl<F, B> fmt::Debug for HandleErrorLayer<F, B> {
     }
 }
 
-/// A [`Service`] adapter that handles errors by converting them into responses.
-///
-/// See [module docs](self) for more details on axum's error handling model.
-pub struct HandleError<S, F, B> {
-    inner: S,
-    f: F,
-    _marker: PhantomData<fn() -> B>,
-}
-
-impl<S, F, B> Clone for HandleError<S, F, B>
+impl<S, F, T> Layer<S> for HandleErrorLayer<F, T>
 where
-    S: Clone,
     F: Clone,
 {
-    fn clone(&self) -> Self {
-        Self::new(self.inner.clone(), self.f.clone())
+    type Service = HandleError<S, F, T>;
+
+    fn layer(&self, inner: S) -> Self::Service {
+        HandleError::new(inner, self.f.clone())
     }
 }
 
-impl<S, F, B> HandleError<S, F, B> {
+/// A [`Service`] adapter that handles errors by converting them into responses.
+///
+/// See [module docs](self) for more details on axum's error handling model.
+pub struct HandleError<S, F, T> {
+    inner: S,
+    f: F,
+    _extractor: PhantomData<fn() -> T>,
+}
+
+impl<S, F, T> HandleError<S, F, T> {
     /// Create a new `HandleError`.
     pub fn new(inner: S, f: F) -> Self {
         Self {
             inner,
             f,
-            _marker: PhantomData,
+            _extractor: PhantomData,
         }
     }
 }
 
-impl<S, F, B> fmt::Debug for HandleError<S, F, B>
+impl<S, F, T> Clone for HandleError<S, F, T>
+where
+    S: Clone,
+    F: Clone,
+{
+    fn clone(&self) -> Self {
+        Self {
+            inner: self.inner.clone(),
+            f: self.f.clone(),
+            _extractor: PhantomData,
+        }
+    }
+}
+
+impl<S, F, E> fmt::Debug for HandleError<S, F, E>
 where
     S: fmt::Debug,
 {
@@ -109,41 +114,114 @@ where
     }
 }
 
-impl<S, F, ReqBody, ResBody, Res> Service<Request<ReqBody>> for HandleError<S, F, ReqBody>
+impl<S, F, ReqBody, ResBody, Fut, Res> Service<Request<ReqBody>> for HandleError<S, F, ()>
 where
-    S: Service<Request<ReqBody>, Response = Response<ResBody>> + Clone,
-    F: FnOnce(S::Error) -> Res + Clone,
+    S: Service<Request<ReqBody>, Response = Response<ResBody>> + Clone + Send + 'static,
+    S::Error: Send,
+    S::Future: Send,
+    F: FnOnce(S::Error) -> Fut + Clone + Send + 'static,
+    Fut: Future<Output = Res> + Send,
     Res: IntoResponse,
-    ResBody: http_body::Body<Data = Bytes> + Send + 'static,
+    ReqBody: Send + 'static,
+    ResBody: HttpBody<Data = Bytes> + Send + 'static,
     ResBody::Error: Into<BoxError>,
 {
     type Response = Response<BoxBody>;
     type Error = Infallible;
-    type Future = future::HandleErrorFuture<Oneshot<S, Request<ReqBody>>, F>;
+    type Future = future::HandleErrorFuture;
 
-    #[inline]
-    fn poll_ready(&mut self, _cx: &mut Context<'_>) -> Poll<Result<(), Self::Error>> {
+    fn poll_ready(&mut self, _: &mut Context<'_>) -> Poll<Result<(), Self::Error>> {
         Poll::Ready(Ok(()))
     }
 
     fn call(&mut self, req: Request<ReqBody>) -> Self::Future {
-        future::HandleErrorFuture {
-            f: Some(self.f.clone()),
-            inner: self.inner.clone().oneshot(req),
+        let f = self.f.clone();
+
+        let clone = self.inner.clone();
+        let inner = std::mem::replace(&mut self.inner, clone);
+
+        let future = Box::pin(async move {
+            match inner.oneshot(req).await {
+                Ok(res) => Ok(res.map(boxed)),
+                Err(err) => Ok(f(err).await.into_response().map(boxed)),
+            }
+        });
+
+        future::HandleErrorFuture { future }
+    }
+}
+
+#[allow(unused_macros)]
+macro_rules! impl_service {
+    ( $($ty:ident),* $(,)? ) => {
+        impl<S, F, ReqBody, ResBody, Res, Fut, $($ty,)*> Service<Request<ReqBody>>
+            for HandleError<S, F, ($($ty,)*)>
+        where
+            S: Service<Request<ReqBody>, Response = Response<ResBody>> + Clone + Send + 'static,
+            S::Error: Send,
+            S::Future: Send,
+            F: FnOnce($($ty),*, S::Error) -> Fut + Clone + Send + 'static,
+            Fut: Future<Output = Res> + Send,
+            Res: IntoResponse,
+            $( $ty: FromRequest<ReqBody> + Send,)*
+            ReqBody: Send + 'static,
+            ResBody: HttpBody<Data = Bytes> + Send + 'static,
+            ResBody::Error: Into<BoxError>,
+        {
+            type Response = Response<BoxBody>;
+            type Error = Infallible;
+
+            type Future = future::HandleErrorFuture;
+
+            fn poll_ready(&mut self, _: &mut Context<'_>) -> Poll<Result<(), Self::Error>> {
+                Poll::Ready(Ok(()))
+            }
+
+            #[allow(non_snake_case)]
+            fn call(&mut self, req: Request<ReqBody>) -> Self::Future {
+                let f = self.f.clone();
+
+                let clone = self.inner.clone();
+                let inner = std::mem::replace(&mut self.inner, clone);
+
+                let future = Box::pin(async move {
+                    let mut req = RequestParts::new(req);
+
+                    $(
+                        let $ty = match $ty::from_request(&mut req).await {
+                            Ok(value) => value,
+                            Err(rejection) => return Ok(rejection.into_response().map(boxed)),
+                        };
+                    )*
+
+                    let req = match req.try_into_request() {
+                        Ok(req) => req,
+                        Err(err) => {
+                            return Ok(Response::builder()
+                                .status(StatusCode::INTERNAL_SERVER_ERROR)
+                                .body(boxed(Full::from(err.to_string())))
+                                .unwrap());
+                        }
+                    };
+
+                    match inner.oneshot(req).await {
+                        Ok(res) => Ok(res.map(boxed)),
+                        Err(err) => Ok(f($($ty),*, err).await.into_response().map(boxed)),
+                    }
+                });
+
+                future::HandleErrorFuture { future }
+            }
         }
     }
 }
 
+all_the_tuples!(impl_service);
+
 pub mod future {
     //! Future types.
 
-    use crate::{
-        body::{boxed, BoxBody},
-        response::IntoResponse,
-        BoxError,
-    };
-    use bytes::Bytes;
-    use futures_util::ready;
+    use crate::body::BoxBody;
     use http::Response;
     use pin_project_lite::pin_project;
     use std::{
@@ -154,36 +232,21 @@ pub mod future {
     };
 
     pin_project! {
-        /// Response future for [`HandleError`](super::HandleError).
-        #[derive(Debug)]
-        pub struct HandleErrorFuture<Fut, F> {
+        /// Response future for [`HandleError`].
+        pub struct HandleErrorFuture {
             #[pin]
-            pub(super) inner: Fut,
-            pub(super) f: Option<F>,
+            pub(super) future: Pin<Box<dyn Future<Output = Result<Response<BoxBody>, Infallible>>
+                + Send
+                + 'static
+            >>,
         }
     }
 
-    impl<Fut, F, E, B, Res> Future for HandleErrorFuture<Fut, F>
-    where
-        Fut: Future<Output = Result<Response<B>, E>>,
-        F: FnOnce(E) -> Res,
-        Res: IntoResponse,
-        B: http_body::Body<Data = Bytes> + Send + 'static,
-        B::Error: Into<BoxError>,
-    {
+    impl Future for HandleErrorFuture {
         type Output = Result<Response<BoxBody>, Infallible>;
 
         fn poll(self: Pin<&mut Self>, cx: &mut Context<'_>) -> Poll<Self::Output> {
-            let this = self.project();
-
-            match ready!(this.inner.poll(cx)) {
-                Ok(res) => Ok(res.map(boxed)).into(),
-                Err(err) => {
-                    let f = this.f.take().unwrap();
-                    let res = f(err);
-                    Ok(res.into_response().map(boxed)).into()
-                }
-            }
+            self.project().future.poll(cx)
         }
     }
 }
