+++ axum/src/extract/tuple.rs
@@ -40,19 +40,4 @@ macro_rules! impl_from_request {
     };
 }
 
-impl_from_request!(T1);
-impl_from_request!(T1, T2);
-impl_from_request!(T1, T2, T3);
-impl_from_request!(T1, T2, T3, T4);
-impl_from_request!(T1, T2, T3, T4, T5);
-impl_from_request!(T1, T2, T3, T4, T5, T6);
-impl_from_request!(T1, T2, T3, T4, T5, T6, T7);
-impl_from_request!(T1, T2, T3, T4, T5, T6, T7, T8);
-impl_from_request!(T1, T2, T3, T4, T5, T6, T7, T8, T9);
-impl_from_request!(T1, T2, T3, T4, T5, T6, T7, T8, T9, T10);
-impl_from_request!(T1, T2, T3, T4, T5, T6, T7, T8, T9, T10, T11);
-impl_from_request!(T1, T2, T3, T4, T5, T6, T7, T8, T9, T10, T11, T12);
-impl_from_request!(T1, T2, T3, T4, T5, T6, T7, T8, T9, T10, T11, T12, T13);
-impl_from_request!(T1, T2, T3, T4, T5, T6, T7, T8, T9, T10, T11, T12, T13, T14);
-impl_from_request!(T1, T2, T3, T4, T5, T6, T7, T8, T9, T10, T11, T12, T13, T14, T15);
-impl_from_request!(T1, T2, T3, T4, T5, T6, T7, T8, T9, T10, T11, T12, T13, T14, T15, T16);
+all_the_tuples!(impl_from_request);
