+++ axum/src/routing/method_routing.rs
@@ -1,6 +1,6 @@
 use crate::{
     body::{boxed, Body, BoxBody, Bytes},
-    error_handling::HandleErrorLayer,
+    error_handling::{HandleError, HandleErrorLayer},
     handler::Handler,
     http::{Method, Request, Response, StatusCode},
     routing::{Fallback, MethodFilter, Route},
@@ -795,12 +795,15 @@ impl<ReqBody, E> MethodRouter<ReqBody, E> {
     /// Apply a [`HandleErrorLayer`].
     ///
     /// This is a convenience method for doing `self.layer(HandleErrorLayer::new(f))`.
-    pub fn handle_error<F, Res>(self, f: F) -> MethodRouter<ReqBody, Infallible>
+    pub fn handle_error<F, T>(self, f: F) -> MethodRouter<ReqBody, Infallible>
     where
-        F: FnOnce(E) -> Res + Clone + Send + 'static,
-        Res: crate::response::IntoResponse,
-        ReqBody: Send + 'static,
+        F: Clone + Send + 'static,
+        HandleError<Route<ReqBody, E>, F, T>:
+            Service<Request<ReqBody>, Response = Response<BoxBody>, Error = Infallible>,
+        <HandleError<Route<ReqBody, E>, F, T> as Service<Request<ReqBody>>>::Future: Send,
+        T: 'static,
         E: 'static,
+        ReqBody: 'static,
     {
         self.layer(HandleErrorLayer::new(f))
     }
@@ -1047,12 +1050,17 @@ mod tests {
             get(ok)
                 .post(ok)
                 .route_layer(RequireAuthorizationLayer::bearer("password"))
-                .merge(delete_service(ServeDir::new(".")).handle_error(|_| StatusCode::NOT_FOUND))
+                .merge(
+                    delete_service(ServeDir::new("."))
+                        .handle_error(|_| async { StatusCode::NOT_FOUND }),
+                )
                 .fallback((|| async { StatusCode::NOT_FOUND }).into_service())
                 .put(ok)
                 .layer(
                     ServiceBuilder::new()
-                        .layer(HandleErrorLayer::new(|_| StatusCode::REQUEST_TIMEOUT))
+                        .layer(HandleErrorLayer::new(|_| async {
+                            StatusCode::REQUEST_TIMEOUT
+                        }))
                         .layer(TimeoutLayer::new(Duration::from_secs(10))),
                 ),
         );
