+++ axum/src/routing/tests/nest.rs
@@ -169,7 +169,7 @@ async fn nested_service_sees_stripped_uri() {
 async fn nest_static_file_server() {
     let app = Router::new().nest(
         "/static",
-        get_service(ServeDir::new(".")).handle_error(|error| {
+        get_service(ServeDir::new(".")).handle_error(|error| async move {
             (
                 StatusCode::INTERNAL_SERVER_ERROR,
                 format!("Unhandled internal error: {}", error),
