+++ examples/sse/src/main.rs
@@ -26,7 +26,7 @@ async fn main() {
 
     let static_files_service =
         get_service(ServeDir::new("examples/sse/assets").append_index_html_on_directories(true))
-            .handle_error(|error: std::io::Error| {
+            .handle_error(|error: std::io::Error| async move {
                 (
                     StatusCode::INTERNAL_SERVER_ERROR,
                     format!("Unhandled internal error: {}", error),
